feat: enrollment use case

+++ src/use-cases/enrollment/destroy-enrollment.ts
@@ -1,4 +1,5 @@
 import { EnrollmentsRepository } from "@/repositories/enrollment-repository"
+import { EnrollmentNotFoundError } from "../errors/enrollment-not-found";
 
 interface DestroyEnrollmentUseCaseRequest {
   enrollmentId: number
@@ -10,6 +11,12 @@ export class DestroyEnrollmentUseCase {
   async execute({
     enrollmentId,
   }: DestroyEnrollmentUseCaseRequest): Promise<Boolean> {
+    let enrollment = await this.enrollmentRepository.checkStatus(enrollmentId)
+
+    if (!enrollment) {
+      throw new EnrollmentNotFoundError();
+    }
+
     return await this.enrollmentRepository.destroy(
       enrollmentId,
     )
